Shift Products section to Home page and migrate Home page content to About page
