+++ app/about/page.jsx
@@ -1,0 +1,10 @@
+import "../../styles/About.css";
+
+export default function Home() {
+  return (
+    <main className="about">
+      <h1>Your Story Starts With Us.</h1>
+      <button className="button">Learn More</button>
+    </main>
+  );
+}
