+++ styles/Blog.css
@@ -24,14 +24,14 @@
 }
 
 .article-image {
-  width: 100%;
-  height: auto;
+  width: auto;
+  height: 100%;
   border-radius: 5px;
 }
 
-/* .article-details {
+.article-details {
   flex: 3;
-} */
+}
 
 .article-title {
   margin-top: 0;
