--- styles/Products.css
@@ -1,33 +1,0 @@
-.store h1 {
-  margin: 20px 0 20px 0;
-  font-size: 3vw;
-}
-
-.products-grid {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 20px;
-  padding: 20px;
-}
-
-.product {
-  border: 1px solid #ccc;
-  padding: 20px;
-  text-align: center;
-}
-
-.product img {
-  max-width: 100%;
-  max-height: 250px;
-  object-fit: contain;
-  margin-bottom: 10px;
-}
-
-.product-price {
-  font-weight: bold;
-  margin: 10px 0;
-}
-
-.sort-button {
-  margin-top: 20px;
-}
