+++ styles/page.css
@@ -1,16 +1,33 @@
-.home {
-  flex-grow: 1;
-  padding: 20px;
-  text-align: center;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  flex-direction: column;
-  height: 100vh;
+.store h1 {
+  margin: 20px 0 20px 0;
+  font-size: 3vw;
 }
 
-.home h1 {
-  margin: 20px 0 20px 0;
-  color: purple;
-  font-size: 3vw;
+.products-grid {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 20px;
+  padding: 20px;
+}
+
+.product {
+  border: 1px solid #ccc;
+  padding: 20px;
+  text-align: center;
+}
+
+.product img {
+  max-width: 100%;
+  max-height: 250px;
+  object-fit: contain;
+  margin-bottom: 10px;
+}
+
+.product-price {
+  font-weight: bold;
+  margin: 10px 0;
+}
+
+.sort-button {
+  margin-top: 20px;
 }
